+++ pelutils/jsonl.py
@@ -6,15 +6,21 @@ import json
 from typing import Generator, Iterable, TextIO
 
 
-def load_jsonl(f: TextIO) -> Generator:
+def load(f: TextIO) -> Generator:
     """ Returns a generator of parsed lines in a .jsonl file. Empty lines are ignored """
     for line in f.readlines():
         stripped = line.strip()
         if stripped:
             yield json.loads(stripped)
 
+def loads(string: str) -> Generator:
+    """ Returns a generator of parsed lines. Empty lines are ignored """
+    for line in string.splitlines():
+        stripped = line.strip()
+        if stripped:
+            yield json.loads(stripped)
 
-def write_jsonl(objects: Iterable, f: TextIO, single_block=True):
+def dump(objects: Iterable, f: TextIO, single_block=True):
     """
     Saves an iterable to a .jsonl file
     If single_block is True, objects will be joined to a single block before being written
@@ -25,3 +31,7 @@ def write_jsonl(objects: Iterable, f: TextIO, single_block=True):
     else:
         for obj in objects:
             f.write(json.dumps(obj) + "\n")
+
+def dumps(objects: Iterable) -> str:
+    """ Returns a string representation of an iterable in a string """
+    return "\n".join(json.dumps(obj) for obj in objects)